# apps/api/openapi/generated.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/samples/{id}": {
      "get": {
        "operationId": "SamplesController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SamplesDto"
                }
              }
            }
          }
        },
        "tags": [
          "samples"
        ]
      },
      "patch": {
        "operationId": "SamplesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SamplesDto"
                }
              }
            }
          }
        },
        "tags": [
          "samples"
        ]
      },
      "delete": {
        "operationId": "SamplesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "samples"
        ]
      }
    },
...
  },
  "info": {
    "title": "AuditForge ISO 42001 API",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "SamplesDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "firmId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "firmId",
          "name",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}